requirements.txt
torch>=2.0.0
transformers>=4.37.0
Pillow>=9.0.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.9.0
accelerate>=0.20.0
bitsandbytes>=0.41.0